+++ src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.hizencode.javafx.main.Launcher
+Main-Class: com.hizencode.excelphoneresolver.main.Launcher
