+++ src/main/java/com/hizencode/excelphoneresolver/main/Launcher.java
@@ -1,4 +1,4 @@
-package com.hizencode.javafx.main;
+package com.hizencode.excelphoneresolver.main;
 
 public class Launcher {
     public static void main(String[] args) {
