+++ src/main/java/module-info.java
@@ -4,6 +4,8 @@ module ExcelPhoneResolver {
     requires org.apache.poi.poi;
     requires org.apache.poi.ooxml;
 
-    opens com.hizencode.javafx.controller to javafx.fxml;
-    opens com.hizencode.javafx.main to javafx.graphics;
+    opens com.hizencode.excelphoneresolver.main to javafx.graphics;
+    opens com.hizencode.excelphoneresolver.ui.start.controller to javafx.fxml;
+    opens com.hizencode.excelphoneresolver.ui.alertmanager to javafx.fxml;
+    opens com.hizencode.excelphoneresolver.ui.resolve.controller to javafx.fxml;
 }
